requirements.txt
beautifulsoup4==4.11.1
Flask==2.2.2
GitPython==3.1.29
html2text==2020.1.16
lxml==4.6.3
pyftpdlib==1.5.7
requests==2.26.0

